test: move legacy-zone tests into test/js/web/intl/intl.test.ts
Per test/CLAUDE.md — add coverage to the existing module test file
rather than a new top-level file. The zones are grouped under
`Intl.DateTimeFormat timeZone option` immediately after the existing
`Intl.DateTimeFormat` describe that covers snapshot + timeZoneName
behaviour. Also switch the macOS guard to `isMacOS` from `harness`
to match the file's existing `isLinux` import.
